Refactor `Datasets/*`
This commit doesn't change any functionality or rendering, however
it contains a suite of (opinionated) renaming and refactoring. It was
motivated by how large `dataset-select.jsx` had become and, at its core,
it simplifies that component by separating out:
1. Dataset filtering UI
2. Filter display UI
3. Listing of filtered datasets.
Within each of these further changes are possible, but are left for
future work.
