+++ static-site/src/components/Datasets/dataset-select.jsx
@@ -1,149 +1,11 @@
 import React from "react";
 import "react-select/dist/react-select.css";
 import "react-virtualized-select/styles.css";
-import Select from "react-virtualized-select";
-import { debounce, get, sortBy } from 'lodash';
-import styled from 'styled-components';
-import ReactTooltip from 'react-tooltip';
-import { FaInfoCircle } from "react-icons/fa";
-import { MdPerson } from "react-icons/md";
-import {Grid, Col, Row} from 'react-styled-flexboxgrid';
-import { FilterBadge, Tooltip } from "./filterBadge";
-import * as splashStyles from "../splash/styles";
-
-const logoPNG = require("../../../static/logos/favicon.png");
-
-const StyledLinkContainer = styled.div`
-  a {
-    color: #444;
-    font-weight: ${(props) => props.bold ? 700 : "normal"};
-  }
-  a:hover,
-  a:focus {
-    color: #5097BA;
-    text-decoration: underline;
-  }
-`;
-
-const StyledIconLinkContainer = styled.div`
-  svg {
-    color: #444;
-  }
-  svg:hover,
-  svg:focus {
-    color: #5097BA;
-  }
-`;
-
-const StyledTooltip = styled(ReactTooltip)`
-  max-width: 30vh;
-  white-space: normal;
-  line-height: 1.2;
-  padding: 10px !important; /* override internal styling */
-  z-index: 1002 !important; /* on top of viz legend */
-  pointer-events: auto !important;
-`;
-
-const Intersect = ({id}) => (
-  <span style={{fontSize: "2rem", fontWeight: 300, padding: "0px 4px 0px 2px", cursor: 'help'}} data-tip data-for={id}>
-    ∩
-    <Tooltip id={id}>{`Groups of filters are combined by intersection`}</Tooltip>
-  </span>
-);
-const Union = () => (
-  <span style={{fontSize: "1.5rem", padding: "0px 3px 0px 2px"}}>
-    ∪
-  </span>
-);
-const openBracketBig = <span style={{fontSize: "2rem", fontWeight: 300, padding: "0px 0px 0px 2px"}}>{'{'}</span>;
-const closeBracketBig = <span style={{fontSize: "2rem", fontWeight: 300, padding: "0px 2px"}}>{'}'}</span>;
-const openBracketSmall = <span style={{fontSize: "1.8rem", fontWeight: 300, padding: "0px 2px"}}>{'{'}</span>;
-const closeBracketSmall = <span style={{fontSize: "1.8rem", fontWeight: 300, padding: "0px 2px"}}>{'}'}</span>;
-
-const DEBOUNCE_TIME = 200;
-
-const DatasetSelectionResultsContainer = styled.div`
-  height: 600px;
-  overflow-x: hidden;
-  overflow-y: visible;
-`;
-
-const DatasetContainer = styled.div`
-  font-family: ${(props) => props.theme.generalFont};
-  font-weight: 900;
-  font-size: 18px;
-  padding: 10px 1px 10px 1px;
-  line-height: 24px;
-`;
-
-const LogoContainer = styled.a`
-  padding: 1px 1px;
-  margin-right: 5px;
-  width: 24px;
-  cursor: pointer;
-`;
-
-const renderDatasets = (datasets, showDates) => {
-  return (
-    <>
-      <Grid fluid>
-        <DatasetContainer key="Column labels" style={{borderBottom: "1px solid #CCC"}}>
-          <Row>
-            <Col xs={8} sm={6} md={7}>
-              Dataset
-            </Col>
-            <Col xs={false} sm={3} md={3}>
-              Contributor
-            </Col>
-            {showDates && <Col xs={false} sm={3} md={2}>
-              Uploaded date
-            </Col>}
-            <Col xs={4} sm={false} style={{textAlign: "right"}}>
-              Contributor
-            </Col>
-          </Row>
-        </DatasetContainer>
-        <DatasetSelectionResultsContainer>
-          { datasets.map((dataset) => (
-            <DatasetContainer key={dataset.filename}>
-              <Row>
-                <Col xs={10} sm={6} md={7}>
-                  <StyledLinkContainer bold>
-                    <a href={dataset.url}>{dataset.filename.replace(/_/g, ' / ').replace('.json', '')}</a>
-                  </StyledLinkContainer>
-                </Col>
-                <Col xs={false} sm={3} md={3}>
-                  <span>
-                    {dataset.contributor.includes("Nextstrain") && <LogoContainer href="https://nextstrain.org">
-                      <img alt="nextstrain.org" className="logo" width="24px" src={logoPNG}/>
-                    </LogoContainer>}
-                    {dataset.contributorUrl === undefined ?
-                      dataset.contributor :
-                      <StyledLinkContainer>
-                        <a href={dataset.contributorUrl}>{dataset.contributor}</a>
-                      </StyledLinkContainer>}
-                  </span>
-                </Col>
-                {showDates && <Col xs={false} sm={3} md={2}>
-                  {dataset.date_uploaded}
-                </Col>}
-                <Col xs={2} sm={false} style={{textAlign: "right"}}>
-                  <LogoContainer href={dataset.contributor.includes("Nextstrain") ? "https://nextstrain.org" : get(dataset, "contributorUrl")}>
-                    {dataset.contributor.includes("Nextstrain") ?
-                      <img alt="nextstrain.org" className="logo" width="24px" src={logoPNG}/> :
-                      <StyledIconLinkContainer><MdPerson/></StyledIconLinkContainer>
-                    }
-                  </LogoContainer>
-                </Col>
-              </Row>
-            </DatasetContainer>
-          ))
-          }
-        </DatasetSelectionResultsContainer>
-      </Grid>
-    </>
-  );
-};
+import { get, sortBy } from 'lodash';
+import { DatasetFilteringSelection } from "./filter-selection";
+import { ListDatasets } from "./list-datasets";
+import { FilterDisplay } from "./filter-display";
+import { computeFilterValues } from "./filter-helpers";
 
 /**
  * <FilterData> is a (keyboard)-typing based search box intended to
@@ -152,73 +14,19 @@ const renderDatasets = (datasets, showDates) => {
  * entries within each category and then take the intersection of those unions.
  */
 class DatasetSelect extends React.Component {
+
   constructor(props) {
     super(props);
     this.state = {
-      filters: {}
+      filters: {},
     };
-    this.applyFilter = this.applyFilter.bind(this);
-    this.createFilterBadges = this.createFilterBadges.bind(this);
-    this.getStyles = this.getStyles.bind(this);
-    this.makeOptions = this.makeOptions.bind(this);
-    this.selectionMade = this.selectionMade.bind(this);
-    this.createIndividualBadge = this.createIndividualBadge.bind(this);
+    this.applyFilter = (mode, trait, values) => {
+      const filters = computeFilterValues(this.state.filters, mode, trait, values);
+      if (filters) this.setState({filters});
+    };
     this.getFilteredDatasets = this.getFilteredDatasets.bind(this);
   }
 
-  applyFilter = (mode, trait, values) => {
-    const currentlyFilteredTraits = Reflect.ownKeys(this.state.filters);
-    let newValues;
-    switch (mode) {
-      case "set":
-        newValues = values.map((value) => ({value, active: true}));
-        break;
-      case "add":
-        if (currentlyFilteredTraits.indexOf(trait) === -1) {
-          newValues = values.map((value) => ({value, active: true}));
-        } else {
-          newValues = this.state.filters[trait].slice();
-          const currentItemNames = newValues.map((i) => i.value);
-          values.forEach((valueToAdd) => {
-            const idx = currentItemNames.indexOf(valueToAdd);
-            if (idx === -1) {
-              newValues.push({value: valueToAdd, active: true});
-            } else {
-              /* it's already there, ensure it's active */
-              newValues[idx].active = true;
-            }
-          });
-        }
-        break;
-      case "remove": // fallthrough
-      case "inactivate":
-        if (currentlyFilteredTraits.indexOf(trait) === -1) {
-          console.error(`trying to ${mode} values from an un-initialised filter!`);
-          return;
-        }
-        newValues = this.state.filters[trait].slice();
-        const currentItemNames = newValues.map((i) => i.value);
-        for (const item of values) {
-          const idx = currentItemNames.indexOf(item);
-          if (idx !== -1) {
-            if (mode==="remove") {
-              newValues.splice(idx, 1);
-            } else {
-              newValues[idx].active = false;
-            }
-          } else {
-            console.error(`trying to ${mode} filter value ${item} which was not part of the filter selection`);
-          }
-        }
-        break;
-      default:
-        console.error(`applyFilter called with invalid mode: ${mode}`);
-        return; // don't set state
-    }
-    const filters = Object.assign({}, this.state.filters, {});
-    filters[trait] = newValues;
-    this.setState({filters});
-  };
   componentDidMount() {
     if (!this.props.urlDefinedFilterPath) return;
     const filterValues = this.props.urlDefinedFilterPath
@@ -227,68 +35,6 @@ class DatasetSelect extends React.Component {
     this.applyFilter("add", "keyword", filterValues);
   }
 
-  getStyles() {
-    return {
-      base: {
-        marginBottom: 0,
-        fontSize: 14
-      }
-    };
-  }
-  makeOptions = () => {
-    /**
-     * The <Select> component needs an array of options to display (and search across). We compute this
-     * by looping across each filter and calculating all valid options for each. This function runs
-     * each time a filter is toggled on / off.
-     */
-    const datasets = this.props.datasets.filter((b) => b.url !== undefined);
-    const optionsObject = datasets.reduce((accumulator, dataset) => {
-      const filename = get(dataset, "filename");
-      const keywordArray = filename.replace('.json', '').split("_");
-      keywordArray.forEach((keyword) => {
-        if (accumulator.seenValues["keyword"]) {
-          if (accumulator.seenValues["keyword"].has(keyword)) return;
-        } else {
-          accumulator.seenValues["keyword"] = new Set([]);
-        }
-        accumulator.seenValues["keyword"].add(keyword);
-        accumulator.options.push({label: `keyword → ${keyword}`, value: ["keyword", keyword]});
-      });
-      return accumulator;
-    }, {options: [], seenValues: {}});
-    return sortBy(optionsObject.options, [(o) => o.label.toLowerCase()]);
-  }
-  selectionMade = (sel) => {
-    this.applyFilter("add", sel.value[0], [sel.value[1]]);
-  }
-  createIndividualBadge({filterName, item, label, onHoverMessage}) {
-    return (
-      <FilterBadge
-        key={item.value}
-        id={String(item.value)}
-        remove={() => {this.applyFilter("remove", filterName, [item.value]);}}
-        canMakeInactive
-        onHoverMessage={onHoverMessage}
-        active={item.active}
-        activate={() => {this.applyFilter("add", filterName, [item.value]);}}
-        inactivate={() => {this.applyFilter("inactivate", filterName, [item.value]);}}
-      >
-        {label}
-      </FilterBadge>
-    );
-  }
-  createFilterBadges(filterName) {
-    const nFilterValues = this.state.filters[filterName].length;
-    const onHoverMessage = nFilterValues === 1 ?
-      `Filtering datasets to this ${filterName}` :
-      `Filtering datasets to these ${nFilterValues} ${filterName}`;
-    return this.state.filters[filterName]
-      .sort((a, b) => a.value < b.value ? -1 : a.value > b.value ? 1 : 0)
-      .map((item) => {
-        const label = `${item.value}`;
-        return this.createIndividualBadge({filterName, item, label, onHoverMessage});
-      });
-  }
   buildMatchesFilter(build, filterName, filterObjects) {
     const keywordArray = get(build, "filename").replace('.json', '').split("_");
     return filterObjects.every((filter) => {
@@ -296,110 +42,35 @@ class DatasetSelect extends React.Component {
       return keywordArray.includes(filter.value);
     });
   }
+
   getFilteredDatasets() {
-    // TODO this doesnt care about categories
+    // TODO this doesn't care about categories
     const filtered = this.props.datasets
       .filter((b) => b.url !== undefined)
       .filter((b) => Object.entries(this.state.filters)
-                      .filter((filterEntry) => filterEntry[1].length)
-                      .every(([filterName, filterValues]) => this.buildMatchesFilter(b, filterName, filterValues)));
+        .filter((filterEntry) => filterEntry[1].length)
+        .every(([filterName, filterValues]) => this.buildMatchesFilter(b, filterName, filterValues)));
     return sortBy(filtered, [(d) => d.filename.toLowerCase()]);
   }
 
   render() {
     // options only need to be calculated a single time per render, and by adding a debounce
-    // to `loadOptions` we don't slow things down by comparing queries to a large number of options
-    const options = this.makeOptions();
-    const loadOptions = debounce((input, callback) => callback(null, {options}), DEBOUNCE_TIME);
-    const styles = this.getStyles();
-    const filtersByCategory = [];
-    Reflect.ownKeys(this.state.filters)
-      .filter((filterName) => this.state.filters[filterName].length > 0)
-      .forEach((filterName) => {
-        filtersByCategory.push({name: filterName, badges: this.createFilterBadges(filterName)});
-      });
-    const filteredDatasets = this.getFilteredDatasets();
-    /* When filter categories were dynamically created (via metadata drag&drop) the `options` here updated but `<Async>`
-    seemed to use a cached version of all values & wouldn't update. Changing the key forces a rerender, but it's not ideal */
-    const divKey = String(Object.keys(this.state.filters).length);
+    // to `loadFilterOptions` we don't slow things down by comparing queries to a large number of options
     return (
       <>
-        <div style={styles.base} key={divKey}>
-          <splashStyles.H3 left>
-            {`Filter datasets `}
-            <>
-              <span style={{cursor: "help"}} data-tip data-for={"build-filter-info"}>
-                <FaInfoCircle/>
-              </span>
-              <StyledTooltip type="dark" effect="solid" id={"build-filter-info"}>
-                <>
-                  {`Use this box to filter the displayed list of datasets based upon filtering criteria.`}
-                  <br/>
-                  {/* TODO do we want to keep this set logic? It's currently broken */}
-                  Data is filtered by forming a union of selected values within each category, and then
-                  taking the intersection between categories (if more than one category is selected).
-                </>
-              </StyledTooltip>
-            </>
-          </splashStyles.H3>
-          <Select
-            async
-            name="filterQueryBox"
-            placeholder="Type filter query here..."
-            value={undefined}
-            arrowRenderer={null}
-            loadOptions={loadOptions}
-            ignoreAccents={false}
-            clearable={false}
-            searchable
-            multi={false}
-            valueKey="label"
-            onChange={this.selectionMade}
-          />
-          {filtersByCategory.length ? (
-            <>
-              {"Filtered to "}
-              {filtersByCategory.map((filterCategory, idx) => {
-                const multipleFilterBadges = filterCategory.badges.length > 1;
-                const previousCategoriesRendered = idx!==0;
-                return (
-                  <span style={{fontSize: "2rem", padding: "0px 2px"}} key={filterCategory.name}>
-                    {previousCategoriesRendered && <Intersect id={'intersect'+idx}/>}
-                    {multipleFilterBadges && openBracketBig} {/* multiple badges => surround with set notation */}
-                    {filterCategory.badges.map((badge, badgeIdx) => {
-                      if (Array.isArray(badge)) { // if `badge` is an array then we wish to render a set-within-a-set
-                        return (
-                          <span key={badge.map((b) => b.props.id).join("")}>
-                            {openBracketSmall}
-                            {badge.map((el, elIdx) => (
-                              <span key={el.props.id}>
-                                {el}
-                                {elIdx!==badge.length-1 && <Union/>}
-                              </span>
-                            ))}
-                            {closeBracketSmall}
-                            {badgeIdx!==filterCategory.badges.length-1 && ", "}
-                          </span>
-                        );
-                      }
-                      return (
-                        <span key={badge.props.id}>
-                          {badge}
-                          {badgeIdx!==filterCategory.badges.length-1 && ", "}
-                        </span>
-                      );
-                    })}
-                    {multipleFilterBadges && closeBracketBig}
-                  </span>
-                );
-              })}
-              {". "}
-            </>
-          ) : null}
-        </div>
-
-        {renderDatasets(filteredDatasets, !this.props.noDates)}
-
+        <DatasetFilteringSelection
+          key={String(Object.keys(this.state.filters).length)}
+          datasets={this.props.datasets}
+          applyFilter={this.applyFilter}
+        />
+        <FilterDisplay
+          filters={this.state.filters}
+          applyFilter={this.applyFilter}
+        />
+        <ListDatasets
+          datasets={this.getFilteredDatasets()}
+          showDates={!this.props.noDates}
+        />
       </>
     );
   }
